admin function to delete guests

+++ app/controllers/sessions_controller.rb
@@ -16,7 +16,7 @@ class SessionsController < ApplicationController
       self.current_user.save
       session[:token] = nil
     end
-    if self.current_user.sample
+    if self.current_user && self.current_user.sample
       self.current_user.destroy
     end
     redirect_to "/"
